backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
python-whois
requests==2.32.3
bs4
